# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenClinic — Electronic Health Record (EHR) & Governance API",
    "description": "Open, modular, and auditable Electronic Health Record (EHR) and Clinical Management API with hybrid access control (RBAC + ACL), multi-tenancy, dynamic capabilities resolution, and enterprise security (Argon2id + in-memory JWT).",
    "version": "1.0.0",
    "contact": {
      "name": "Iniciativa OpenClinic",
      "url": "https://github.com/Iniciativa-OpenClinic/OpenClinic"
    },
    "license": {
      "name": "GNU AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token in Authorization header: `Bearer <token>`"
      }
    },
    "schemas": {
...
      "def-0": {
        "type": "object",
        "description": "Estrutura padronizada de erros da aplicação no padrão RFC 7807",
        "properties": {
          "type": {
            "type": "string",
            "description": "URI identificadora do tipo de erro"
          },
          "title": {
            "type": "string",
            "description": "Resumo legível do erro"
          },
          "status": {
            "type": "integer",
            "description": "Código de status HTTP"
          },
          "detail": {
            "type": "string",
            "description": "Descrição detalhada e traduzida do erro"
          },
          "code": {
            "type": "string",
            "description": "Código unívoco de erro da aplicação (ex: ERR_AUTH_INVALID_CREDENTIALS)"
          },
          "instance": {
            "type": "string",
            "description": "URI do endpoint onde ocorreu o erro"
          },
          "invalid_params": {
            "type": "array",
            "description": "Lista de parâmetros com falha de validação",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "reason": {
                  "type": "string"
                }
              }
            }
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetails"
      }
    }
  },
  "paths": {
...
    "/api/v1/iam/permissions": {
      "get": {
        "summary": "Get Consolidated User Permission Keys",
        "tags": [
          "IAM & Access Control"
        ],
        "description": "Returns the active set of direct and group-inherited permission keys for the authenticated user.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of permission keys",
            "content": {
              "application/json": {
                "schema": {
                  "description": "List of permission keys",
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "users:view"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida ou erro de validação (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou token inválido/expirado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado — Permissão ou Role insuficiente (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/iam/capabilities": {
      "get": {
        "summary": "Get User Capabilities Matrix",
        "tags": [
          "IAM & Access Control"
        ],
        "description": "Returns the consolidated matrix of authorized resources, routes, permitted actions, and UI metadata.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Matrix of user capabilities",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Matrix of user capabilities",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "key": {
                        "type": "string",
                        "example": "op_schedule"
                      },
                      "label": {
                        "type": "string",
                        "example": "Agenda & Marcações"
                      },
                      "icon": {
                        "type": "string",
                        "nullable": true
                      },
                      "route": {
                        "type": "string",
                        "nullable": true
                      },
                      "resource_type": {
                        "type": "string"
                      },
                      "context": {
                        "type": "string"
                      },
                      "actions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "required": [
                      "key",
                      "label",
                      "resource_type",
                      "context",
                      "actions"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida ou erro de validação (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou token inválido/expirado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado — Permissão ou Role insuficiente (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/iam/navigation": {
      "get": {
        "summary": "Get Navigation Menu by Architecture Context",
        "tags": [
          "IAM & Access Control"
        ],
        "description": "Returns authorized navigation items filtered by application context (ARCH or BUSINESS).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "BUSINESS",
                "ARCH"
              ],
              "default": "BUSINESS"
            },
            "in": "query",
            "name": "context",
            "required": false
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered navigation menu tree",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Filtered navigation menu tree",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "context": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "items": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "item_code": {
                              "type": "string"
                            },
                            "label": {
                              "type": "string"
                            },
                            "icon": {
                              "type": "string",
                              "nullable": true
                            },
                            "route": {
                              "type": "string",
                              "nullable": true
                            },
                            "sort_order": {
                              "type": "number"
                            },
                            "min_role": {
                              "type": "string"
                            },
                            "context": {
                              "type": "string"
                            },
                            "children": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "additionalProperties": true
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida ou erro de validação (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou token inválido/expirado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado — Permissão ou Role insuficiente (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/iam/resources": {
      "get": {
        "summary": "List All System Resources",
        "tags": [
          "IAM & Access Control"
        ],
        "description": "Returns the flat list of all manageable resources within the IAM subsystem.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of application resources",
            "content": {
              "application/json": {
                "schema": {
                  "description": "List of application resources",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "item_code": {
                        "type": "string"
                      },
                      "parent_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                      },
                      "resource_type": {
                        "type": "string"
                      },
                      "context": {
                        "type": "string"
                      },
                      "label_key": {
                        "type": "string"
                      },
                      "route": {
                        "type": "string",
                        "nullable": true
                      },
                      "icon": {
                        "type": "string",
                        "nullable": true
                      },
                      "sort_order": {
                        "type": "number",
                        "nullable": true
                      },
                      "min_role": {
                        "type": "string",
                        "nullable": true
                      },
                      "description": {
                        "type": "string",
                        "nullable": true
                      },
                      "is_active": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida ou erro de validação (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou token inválido/expirado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado — Permissão ou Role insuficiente (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/iam/permissions/group/{id}": {
      "get": {
        "summary": "Get Group ACL Permissions",
        "tags": [
          "IAM & Access Control"
        ],
        "description": "Returns permissions configured for a specific group.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Permissions for the group",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Permissions for the group",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "resource_key": {
                        "type": "string"
                      },
                      "actions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "effect": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida ou erro de validação (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou token inválido/expirado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado — Permissão ou Role insuficiente (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
}